paineluniversal/paineluniversal/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
python-dotenv==1.0.0
alembic==1.12.1
redis==5.0.1
websockets==12.0
httpx==0.25.2
reportlab==4.0.7
pandas==2.1.3
qrcode==7.4.2
pillow==10.1.0
openpyxl==3.1.2
schedule==1.2.0
aiohttp==3.9.1
email-validator==2.1.0